requirements.txt
pyyaml >= 6.0.1
validators >= 0.22.0